:truck: refactor: rename AgentModal to AgentDetail
