+++ src/app/market/features/AgentDetail/AgentInfo/style.ts
